Feature/116

+++ app/src/main/java/com/moritoui/vegegrowthapp/ui/home/HomeScreenViewModel.kt
@@ -2,6 +2,7 @@ package com.moritoui.vegegrowthapp.ui.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.moritoui.vegegrowthapp.data.room.model.VegetableFolderEntity
 import com.moritoui.vegegrowthapp.model.FilterStatus
 import com.moritoui.vegegrowthapp.model.SelectMenu
 import com.moritoui.vegegrowthapp.model.VegeCategory
@@ -10,18 +11,27 @@ import com.moritoui.vegegrowthapp.model.VegeItemDetail
 import com.moritoui.vegegrowthapp.model.VegeStatus
 import com.moritoui.vegegrowthapp.model.filterStatusMap
 import com.moritoui.vegegrowthapp.repository.datamigration.DataMigrationRepository
+import com.moritoui.vegegrowthapp.ui.home.model.AddDialogType
 import com.moritoui.vegegrowthapp.ui.home.model.HomeScreenUiState
 import com.moritoui.vegegrowthapp.ui.home.model.HomeVegetablesState
 import com.moritoui.vegegrowthapp.usecases.AddVegeItemUseCase
 import com.moritoui.vegegrowthapp.usecases.ChangeVegeItemStatusUseCase
+import com.moritoui.vegegrowthapp.usecases.DeleteVegeFolderUseCase
 import com.moritoui.vegegrowthapp.usecases.DeleteVegeItemUseCase
 import com.moritoui.vegegrowthapp.usecases.GetVegeItemDetailLastUseCase
 import com.moritoui.vegegrowthapp.usecases.GetVegeItemListUseCase
+import com.moritoui.vegegrowthapp.usecases.GetVegetableFolderUseCase
+import com.moritoui.vegegrowthapp.usecases.InsertVegetableFolderUseCase
+import com.moritoui.vegegrowthapp.usecases.UpdateVegetableFolderUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import java.util.UUID
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.launchIn
@@ -35,10 +45,14 @@ import javax.inject.Inject
 class HomeScreenViewModel @Inject constructor(
     private val addVegeItemUseCase: AddVegeItemUseCase,
     private val deleteVegeItemUseCase: DeleteVegeItemUseCase,
+    private val deleteVegeFolderUseCase: DeleteVegeFolderUseCase,
     private val getVegeItemListUseCase: GetVegeItemListUseCase,
     private val changeVegeItemStatusUseCase: ChangeVegeItemStatusUseCase,
     private val getVegeItemDetailLast: GetVegeItemDetailLastUseCase,
     private val dataMigrationRepository: DataMigrationRepository,
+    private val getVegetableFolderUseCase: GetVegetableFolderUseCase,
+    private val insertVegetableFolderUseCase: InsertVegetableFolderUseCase,
+    private val updateVegetableFolderUseCase: UpdateVegetableFolderUseCase,
 ) : ViewModel() {
     private val _uiState = MutableStateFlow(HomeScreenUiState.initialState())
     val uiState: StateFlow<HomeScreenUiState> = _uiState.asStateFlow()
@@ -48,19 +62,28 @@ class HomeScreenViewModel @Inject constructor(
 
     private val _vegetables = MutableStateFlow<List<VegeItem>>(emptyList())
     private val _vegetableDetails = MutableStateFlow<List<VegeItemDetail?>>(emptyList())
+    private val _vegetableFolders = MutableStateFlow<List<VegetableFolderEntity>>(emptyList())
     val vegetablesState: StateFlow<HomeVegetablesState> = combine(
         _vegetables,
-        _vegetableDetails
-    ) { vegetables, vegetableDetails ->
+        _vegetableDetails,
+        _vegetableFolders
+    ) { vegetables, vegetableDetails, vegetableFolders ->
         HomeVegetablesState(
             vegetables = vegetables,
-            vegetableDetails = vegetableDetails
+            vegetableDetails = vegetableDetails,
+            vegetableFolders = vegetableFolders
         )
     }.stateIn(
         viewModelScope,
         SharingStarted.WhileSubscribed(5000),
         HomeVegetablesState.initial()
     )
+
+    /**
+     * インサートエラーかどうか
+     */
+    private val _insertVegetableFolderEvent: MutableSharedFlow<Boolean> = MutableStateFlow(false)
+    val insertVegetableFolderEvent: SharedFlow<Boolean> = _insertVegetableFolderEvent.asSharedFlow()
 
     init {
         viewModelScope.launch {
@@ -72,15 +95,15 @@ class HomeScreenViewModel @Inject constructor(
     fun closeDialog() {
         _uiState.update {
             it.copy(
-                isOpenAddDialog = false
+                openAddDialogType = AddDialogType.NotOpenDialog
             )
         }
     }
 
-    fun openAddDialog() {
+    fun openAddDialog(addDialogType: AddDialogType) {
         _uiState.update {
             it.copy(
-                isOpenAddDialog = true,
+                openAddDialogType = addDialogType,
                 inputText = "",
                 selectCategory = VegeCategory.None
             )
@@ -111,18 +134,48 @@ class HomeScreenViewModel @Inject constructor(
         }
     }
 
-    fun saveVegeItem() {
-        val vegeItem =
-            VegeItem(
-                name = _uiState.value.inputText,
-                category = _uiState.value.selectCategory,
-                uuid = UUID.randomUUID().toString(),
-                status = VegeStatus.Default
-            )
-        viewModelScope.launch {
-            addVegeItemUseCase(vegeItem)
-            reloadVegetables()
-            closeDialog()
+    /**
+     * 押されたダイアログの種類によって処理を変える
+     */
+    fun onAddDialogConfirm(addDialogType: AddDialogType) {
+        when (addDialogType) {
+            AddDialogType.AddVegeItem -> {
+                val vegeItem =
+                    VegeItem(
+                        name = _uiState.value.inputText,
+                        category = _uiState.value.selectCategory,
+                        uuid = UUID.randomUUID().toString(),
+                        status = VegeStatus.Default,
+                        folderId = null
+                    )
+                viewModelScope.launch {
+                    addVegeItemUseCase(vegeItem)
+                    reloadVegetables()
+                    closeDialog()
+                }
+            }
+            AddDialogType.AddFolder -> {
+                val vegeFolder = VegetableFolderEntity(
+                    id = 0,
+                    folderNumber = _vegetableFolders.value.size,
+                    folderName = _uiState.value.inputText,
+                    vegetableCategory = _uiState.value.selectCategory
+                )
+                viewModelScope.launch {
+                    insertVegetableFolderUseCase(vegeFolder)
+                        .onSuccess {
+                            reloadVegetables()
+                            closeDialog()
+                        }.onFailure {
+                            _insertVegetableFolderEvent.emit(true)
+                            delay(2000)
+                            _insertVegetableFolderEvent.emit(false)
+                        }
+                }
+            }
+            else -> {
+                return
+            }
         }
     }
 
@@ -153,11 +206,20 @@ class HomeScreenViewModel @Inject constructor(
     }
 
     fun deleteItem() {
-        val deleteItem = _uiState.value.targetDeleteItem ?: return
         viewModelScope.launch {
-            deleteVegeItemUseCase(deleteItem)
+            val deleteItem = _uiState.value.targetDeleteItem
+            val deleteFolder = _uiState.value.targetDeleteFolder
+            if (deleteItem != null) {
+                deleteVegeItemUseCase(deleteItem)
+            }
+            if (deleteFolder != null) {
+                deleteVegeFolderUseCase(deleteFolder)
+            }
             _uiState.update {
-                it.copy(targetDeleteItem = null)
+                it.copy(
+                    targetDeleteItem = null,
+                    targetDeleteFolder = null
+                )
             }
         }
     }
@@ -201,11 +263,23 @@ class HomeScreenViewModel @Inject constructor(
     /**
      * 削除ダイアログを開いて，削除するものをセットする
      */
-    fun openDeleteDialog(vegeItem: VegeItem) {
+    fun openDeleteVegeItemDialog(vegeItem: VegeItem) {
         _uiState.update {
             it.copy(
                 isOpenDeleteDialog = true,
                 targetDeleteItem = vegeItem
+            )
+        }
+    }
+
+    /**
+     * 削除ダイアログを開いて，削除するものをセットする
+     */
+    fun openDeleteFolderDialog(folder: VegetableFolderEntity) {
+        _uiState.update {
+            it.copy(
+                isOpenDeleteDialog = true,
+                targetDeleteFolder = folder
             )
         }
     }
@@ -236,6 +310,10 @@ class HomeScreenViewModel @Inject constructor(
 
             _vegetableDetails.update {
                 filteredVegetables.map { reloadVegetableDetailLast(it) }
+            }
+
+            _vegetableFolders.update {
+                getVegetableFolderUseCase()
             }
         }
     }
